feat(researcher): use login key as MemWal delegate key and remove memory UI
Move ed25519 key derivation server-side so the private key submitted at
login is stored in the JWT session and used directly for MemWal operations
(sprint save, recall, prepare). Remove the client-side MemWal toggle
button, key input dropdown, and all related state/props/localStorage
plumbing from the chat UI.
