+++ apps/researcher/app/(chat)/api/chat/route.ts
@@ -148,7 +148,7 @@ export async function POST(request: Request) {
     const modelMessages = await convertToModelMessages(uiMessages);
 
     // Resolve sprint context — pre-built during preparation and stored on chat record
-    const memwalKey = (requestBody as any).memwalKey || process.env.MEMWAL_KEY;
+    const memwalKey = session.user.privateKey || process.env.MEMWAL_KEY;
     const resolvedSprintIds: string[] = chat?.sprintIds ?? [];
     const prebuiltSprintContext: string | null = chat?.sprintContext ?? null;
